+++ model/lib/model.dart
@@ -2,6 +2,8 @@
 /// only for implementation.
 library model;
 
+export 'src/exception/content_type_mismatched.dart';
+export 'src/exception/non_http_url.dart';
 export 'src/model/metainfo.dart';
 export 'src/parser/open_graph.dart';
 export 'src/parser/twitter_card.dart';