+++ model/lib/src/disguise_ua/disguise_ua.dart
@@ -1,0 +1,2 @@
+export 'disguise_ua_other.dart'
+  if (dart.library.html) 'disguise_ua_web.dart';