+++ model/lib/src/disguise_ua/disguise_ua_other.dart
@@ -1,0 +1,6 @@
+import 'package:meta/meta.dart';
+
+import '../client.dart';
+
+@internal
+String get disguisedUserAgent => OgHrefClient.userAgent;