+++ model/lib/src/disguise_ua/disguise_ua_web.dart
@@ -1,0 +1,6 @@
+import 'dart:html';
+
+import 'package:meta/meta.dart';
+
+@internal
+String get disguisedUserAgent => window.navigator.userAgent;