+++ model/lib/src/exception/non_http_url.dart
@@ -3,6 +3,7 @@ final class NonHttpUrlException implements Exception {
   /// Non-HTTP(S) [Uri].
   final Uri url;
 
+  /// Create an exception which using invalid protocol [url].
   NonHttpUrlException(this.url);
 
   @override